Refactor import paths and remove obsolete components for improved code organization
- Updated import paths for ClusterProvider and related components to use a consistent naming convention, enhancing clarity and maintainability.
- Removed the obsolete `black-scholes-model` and `cluster-data-access` components, streamlining the codebase and reducing complexity.
- Consolidated utility imports and adjusted related components to ensure a cohesive structure across the application.
