+++ src/components/solana/user-settings/rpc-dropdown-select.tsx
@@ -4,8 +4,8 @@ import { useConnection } from '@solana/wallet-adapter-react'
 import { IconTrash } from '@tabler/icons-react'
 import { useQuery } from '@tanstack/react-query'
 import { ReactNode, useState } from 'react'
-import { AppModal } from '../ui/ui-layout'
-import { ClusterNetwork, useCluster } from '../../solana/clusters/cluster-data-access'
+import { AppModal } from '../../ui/ui-layout'
+import { ClusterNetwork, useCluster } from '../../../solana/clusters/clusterDataAccess'
 import { createSolanaClient } from 'gill'
 import {
   DropdownMenu,
