+++ src/components/ui/ui-layout.tsx
@@ -5,8 +5,8 @@ import * as React from 'react'
 import {ReactNode, Suspense, useEffect, useRef} from 'react'
 import toast, {Toaster} from 'react-hot-toast'
 import { useTheme } from 'next-themes'
-import { useCluster } from '../../solana/clusters/cluster-data-access'
-import {ClusterChecker, ExplorerLink} from '../solana/rpc-dropdown-select'
+import { useCluster } from '../../solana/clusters/clusterDataAccess'
+import {ClusterChecker, ExplorerLink} from '../solana/user-settings/rpc-dropdown-select'
 import {WalletBalanceLogger} from '../solana/user-wallet/wallet-balance'
 import { Navbar } from './navbar'
