+++ src/lib/option-pricing-model/blackScholesModel.ts
