--- src/types/mint/form.ts
@@ -1,29 +1,0 @@
-import * as z from "zod";
-import { TOKENS } from "@/constants/token-list/token-list";
-
-export const formSchema = z.object({
-
-  asset: z.enum(Object.keys(TOKENS) as [string, ...string[]]),
-  optionType: z.enum(["call", "put"]),
-  expirationDate: z.date({
-    required_error: "Expiration date is required",
-  }),
-  strikePrice: z.string().refine(val => val !== '', {
-    message: "Strike price is required",
-  }),
-  premium: z.string().refine(
-    (val) => {
-      if (val === '') return true;
-      const num = Number(val);
-      return !isNaN(num) && num >= 0;
-    },
-    { message: "Premium must be a valid number" }
-  ),
-  quantity: z.coerce
-    .number()
-    .int({ message: "Quantity must be a whole number" })
-    .min(1, { message: "Quantity must be at least 1" })
-    .max(100, { message: "Quantity must be at most 100" })
-});
-
-export type FormData = z.infer<typeof formSchema>;
